requirements.txt
Django==2.2.6
django-cors-headers==3.1.1
django-suit==0.2.28
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
psycopg2-binary==2.8.3
python-dotenv==0.10.3
django-suit==0.2.28
coreapi==2.3.3